update timestamp parser in issue tag
currently, we use timestamp tag in issue tag to get the timestamp and
convert it to datetime. but we found in test, that it is possible that
timestamp tag is missing, then raise error. in order to fix it, we use
another tag which are `year`, `month`, `day`, `hour`, `minute`, and
`second`.  these tags contain exactly the same value in timestamp tag.
eventhough it makes the parsing more complex, it allow the flexible API
wrapper.